hello-spring/src/main/java/OneStep/hellospring/repository/MemoryMemberRepository.java
package OneStep.hellospring.repository;

import OneStep.hellospring.domain.Member;

import java.util.*;

public class MemoryMemberRepository implements MemberRepository{

    private static Map<Long, Member> store = new HashMap<>();  // key, Member obj 형태로 저장
    private static long sequence = 0L;  // key 생성하기 위해 선언

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id));  // null 이 나올 수 있으므로 Optional로 감싼다
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream()
                .filter(member -> member.getName().equals(name))
                .findAny();
        // store에서 하나씩 돌면서 name 같은 것 찾음
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }


    public void clearStore(){
        store.clear();
    }
}
